AtomPair related wrapper code moved to its own file
